refactor: add github provider and update discord provider
